requirements.txt
opencv-python==4.3.0.36
opencv-contrib-python==4.3.0.36
numpy
Pillow
imutils
dlib
scipy
